feat(appService): moved appLeadImage endpoint to AppChange controller
Ref: CPLP-2645
