+++ src/marketplace/Apps.Service/BusinessLogic/IAppChangeBusinessLogic.cs
@@ -51,4 +51,13 @@ public interface IAppChangeBusinessLogic
     /// <param name="iamUserId">Id of the iamUser</param>
     /// <param name="offerDescriptionDatas">OfferDescription Data</param>
     Task CreateOrUpdateAppDescriptionByIdAsync(Guid appId, string iamUserId, IEnumerable<LocalizedDescription> offerDescriptionDatas);
+
+    /// <summary>
+    /// Upload OfferAssigned AppLeadImage Document by appId
+    /// </summary>
+    /// <param name="appId">Id of the app</param>
+    /// <param name="iamUserId">Id of the iamUser</param>
+    /// <param name="document">Document Data</param>
+    /// <param name="cancellationToken">cancellationToken</param>
+    Task UploadOfferAssignedAppLeadImageDocumentByIdAsync(Guid appId, string iamUserId, IFormFile document, CancellationToken cancellationToken);
 }
